chore: change return type of AR usecase to return pagingData
