+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/media/ConversationAssetMessagesViewState.kt
@@ -25,14 +25,12 @@ import androidx.paging.PagingData
 import com.wire.android.ui.home.conversations.model.UIMessage
 import com.wire.android.ui.home.conversations.model.messagetypes.asset.UIAssetMessage
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.immutableMapOf
 import kotlinx.collections.immutable.persistentListOf
-import kotlinx.collections.immutable.persistentMapOf
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
 
 @Stable
 data class ConversationAssetMessagesViewState(
     val imageMessages: ImmutableList<UIAssetMessage> = persistentListOf(),
-    val assetMessages: Map<String, List<UIMessage>> = persistentMapOf()
+    val assetMessages: Flow<PagingData<UIMessage>> = emptyFlow()
 )
