+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/usecase/GetAssetMessagesFromConversationUseCase.kt
@@ -18,14 +18,12 @@
 package com.wire.android.ui.home.conversations.usecase
 
 import androidx.paging.PagingConfig
+import androidx.paging.PagingData
 import androidx.paging.flatMap
-import com.wire.android.appLogger
 import com.wire.android.mapper.MessageMapper
-import com.wire.android.ui.common.toGenericAssetGroupedByMonthAndYear
 import com.wire.android.ui.home.conversations.model.UIMessage
 import com.wire.android.util.dispatchers.DispatcherProvider
 import com.wire.kalium.logic.data.id.ConversationId
-import com.wire.kalium.logic.data.message.Message
 import com.wire.kalium.logic.feature.asset.GetPaginatedFlowOfAssetMessageByConversationIdUseCase
 import com.wire.kalium.logic.feature.conversation.ObserveUserListByIdUseCase
 import kotlinx.coroutines.flow.Flow
@@ -33,7 +31,6 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapLatest
-import kotlinx.datetime.TimeZone
 import javax.inject.Inject
 import kotlin.math.max
 
@@ -48,14 +45,12 @@ class GetAssetMessagesFromConversationUseCase @Inject constructor(
      * This operation combines asset messages from a conversation and its respective user to UI
      * @param conversationId The conversation ID that it will look for asset messages in.
      *
-     * String stands for the Month/Year header, and List<UIMessage> of messages to be displayed.
-     * @return A [Map<String, List<UIMessage>>] indicating the success of the operation.
-     *
+     * @return A [PagingData<UIMessage>>] indicating the success of the operation.
      */
     suspend operator fun invoke(
         conversationId: ConversationId,
         initialOffset: Int
-    ): Flow<Map<String, List<UIMessage>>> {
+    ): Flow<PagingData<UIMessage>> {
         val pagingConfig = PagingConfig(
             pageSize = PAGE_SIZE,
             prefetchDistance = PREFETCH_DISTANCE,
@@ -67,75 +62,14 @@ class GetAssetMessagesFromConversationUseCase @Inject constructor(
             startingOffset = max(0, initialOffset - PREFETCH_DISTANCE).toLong(),
             pagingConfig = pagingConfig
         ).map { pagingData ->
-            val messagesList = mutableListOf<Message.Standalone>()
             pagingData.flatMap { messageItem ->
-                appLogger.d("ASSET_FILES -> msg : ${messageItem.conversationId}")
-                messagesList.add(messageItem)
-                listOf(messageItem)
+                observeMemberDetailsByIds(messageMapper.memberIdList(listOf(messageItem)))
+                    .mapLatest { usersList ->
+                        messageMapper.toUIMessage(usersList, messageItem)?.let { listOf(it) }
+                            ?: emptyList()
+                    }.first()
             }
-
-            val timeZone = TimeZone.currentSystemDefault()
-            appLogger.d("ASSET_FILES -> SIZE : ${messagesList.size}")
-            messagesList.toGenericAssetGroupedByMonthAndYear(timeZone = timeZone)
-        }.map { groupedAssetMessages ->
-            mutableMapOf<String, List<UIMessage>>().apply {
-                groupedAssetMessages.entries.forEach { entry ->
-                    val values: List<UIMessage> = entry.value.map { message ->
-                        observeMemberDetailsByIds(messageMapper.memberIdList(listOf(message)))
-                            .mapLatest { usersList ->
-                                messageMapper.toUIMessage(usersList, message)?.let { listOf(it) }
-                                    ?: emptyList()
-                            }.first().first()
-                    }
-                    this@apply.put(entry.key, values)
-                }
-            }
-
-//            val x: String = groupedAssetMessages.entries.map {
-//                val values: List<UIMessage> = it.value.map { message ->
-//                    observeMemberDetailsByIds(messageMapper.memberIdList(listOf(message)))
-//                        .mapLatest { usersList ->
-//                            messageMapper.toUIMessage(usersList, message)?.let { listOf(it) }
-//                                ?: emptyList()
-//                        }
-//                        .first()
-//                }
-//                mapOf(it.key to values)
-            // }
-            // mapOf<String, List<UIMessage>>()
         }.flowOn(dispatchers.io())
-
-//            .map {
-////            pagingData.flatMap { messageItem ->
-////                observeMemberDetailsByIds(messageMapper.memberIdList(listOf(messageItem)))
-////                    .mapLatest { usersList ->
-////                        messageMapper.toUIMessage(usersList, messageItem)?.let { listOf(it) }
-////                            ?: emptyList()
-////                    }.first()
-////            }
-//
-//        }.asFlow().flowOn(dispatchers.io())
-
-//        return getAssetMessages(
-//            conversationId = conversationId,
-//            limit = 35,
-//            offset = 0
-//        ).let { assetMessageList ->
-//            val timeZone = TimeZone.currentSystemDefault()
-//            assetMessageList.toGenericAssetGroupedByMonthAndYear(timeZone = timeZone)
-//        }.map {
-//
-//            val values: List<UIMessage> = it.value.map { message ->
-//                observeMemberDetailsByIds(messageMapper.memberIdList(listOf(message)))
-//                    .mapLatest { usersList ->
-//                        messageMapper.toUIMessage(usersList, message)?.let { listOf(it) }
-//                            ?: emptyList()
-//                    }
-//                    .first()
-//                    .first()
-//            }
-//            mapOf(it.key to values)
-//        }.first()
     }
 
     private companion object {
